Improves/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
Pillow==10.0.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
scikit-learn==1.3.0
dlib==19.24.2
